refactor: 작성한 Todo 관련 컴포넌트를 Todo 페이지에 반영 / catch문의 console.log를 console.error로 변경
